fillers/4844: coalesce utilities into a single file
