+++ tests/eips/eip4844/test_excess_data_gas_fork_transition.py
@@ -19,91 +19,19 @@ from ethereum_test_tools import (
     to_hash_bytes,
 )
 
+from .utils import BLOB_COMMITMENT_VERSION_KZG, MAX_BLOBS_PER_BLOCK, TARGET_BLOBS_PER_BLOCK
+
 REFERENCE_SPEC_GIT_PATH = "EIPS/eip-4844.md"
 REFERENCE_SPEC_VERSION = "ac003985b9be74ff48bd897770e6d5f2e4318715"
 
 # All tests run on the transition fork from Shanghai to Cancun
 pytestmark = pytest.mark.valid_at_transition_to("Cancun")
 
-BLOB_COMMITMENT_VERSION_KZG = 1
-BLOBHASH_GAS_COST = 3
-MIN_DATA_GASPRICE = 1
-DATA_GAS_PER_BLOB = 2**17
-MAX_DATA_GAS_PER_BLOCK = 2**19
-TARGET_DATA_GAS_PER_BLOCK = 2**18
-MAX_BLOBS_PER_BLOCK = MAX_DATA_GAS_PER_BLOCK // DATA_GAS_PER_BLOB
-TARGET_BLOBS_PER_BLOCK = TARGET_DATA_GAS_PER_BLOCK // DATA_GAS_PER_BLOB
-DATA_GASPRICE_UPDATE_FRACTION = 2225652
-
-
-def fake_exponential(factor: int, numerator: int, denominator: int) -> int:
-    """
-    Used to calculate the data gas cost.
-    """
-    i = 1
-    output = 0
-    numerator_accumulator = factor * denominator
-    while numerator_accumulator > 0:
-        output += numerator_accumulator
-        numerator_accumulator = (numerator_accumulator * numerator) // (denominator * i)
-        i += 1
-    return output // denominator
-
-
-def calc_data_fee(tx: Transaction, excess_data_gas: int) -> int:
-    """
-    Calculate the data fee for a transaction.
-    """
-    return get_total_data_gas(tx) * get_data_gasprice(excess_data_gas)
-
-
-def get_total_data_gas(tx: Transaction) -> int:
-    """
-    Calculate the total data gas for a transaction.
-    """
-    if tx.blob_versioned_hashes is None:
-        return 0
-    return DATA_GAS_PER_BLOB * len(tx.blob_versioned_hashes)
-
-
-def get_data_gasprice_from_blobs(excess_blobs: int) -> int:
-    """
-    Calculate the data gas price from the excess blob count.
-    """
-    return fake_exponential(
-        MIN_DATA_GASPRICE,
-        excess_blobs * DATA_GAS_PER_BLOB,
-        DATA_GASPRICE_UPDATE_FRACTION,
-    )
-
-
-def get_data_gasprice(excess_data_gas: int) -> int:
-    """
-    Calculate the data gas price from the excess.
-    """
-    return fake_exponential(
-        MIN_DATA_GASPRICE,
-        excess_data_gas,
-        DATA_GASPRICE_UPDATE_FRACTION,
-    )
-
 
 # Timestamp of the fork
 FORK_TIMESTAMP = 15_000
 # Data gas cost increases from 1 to 2 at this amount of excess blobs
 BLOBS_TO_DATA_GAS_COST_INCREASE = 12
-
-
-def calc_excess_data_gas(parent_excess_data_gas: int, new_blobs: int) -> int:
-    """
-    Calculate the excess data gas for a block given the parent excess data gas
-    and the number of blobs in the block.
-    """
-    consumed_data_gas = new_blobs * DATA_GAS_PER_BLOB
-    if parent_excess_data_gas + consumed_data_gas < TARGET_DATA_GAS_PER_BLOCK:
-        return 0
-    else:
-        return parent_excess_data_gas + consumed_data_gas - TARGET_DATA_GAS_PER_BLOCK
 
 
 @pytest.fixture
